+++ VectorComparisonMapper.java
@@ -6,13 +6,13 @@ import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
-public class VectorComparisonMapper extends Mapper<LongWritable, Text, IntWritable, IntWritable> {
+public class VectorComparisonMapper extends Mapper<LongWritable, Text, IntWritable, DoubleWritable> {
 
 	@Override
 	public void map(LongWritable key, Text value, Context context) throws IOException, InterruptedException {
 		String line = value.toString(); 
 		String[] extracts = line.split(",");
-		context.write(new IntWritable(Integer.parseInt(extracts[1].trim())), new IntWritable(Integer.parseInt(extracts[3].trim())));
+		context.write(new IntWritable(Integer.parseInt(extracts[1].trim())), new DoubleWritable(Double.parseDouble(extracts[3].trim())));
 	}
 }
